chore: group temp model-clean hack in one place
